requirements.txt
numpy
opencv-python==3.4.8.29
pandas
scipy
tqdm
matplotlib
pillow
pyqt5
sklearn
scikit-image
dlib==19.6.1
jupyter
tensorflow